exceptions

+++ src/api.py
@@ -11,14 +11,17 @@ db = NotesDB()
 TIME_FORMAT = '%Y-%m-%d'
 
 
+def get_note(note_id):
+    reply = db.get_note_by_id(note_id)
+    return Note(reply[0], reply[1], reply[2], reply[3].strftime(TIME_FORMAT), reply[4]).__dict__
+
+
 @app.post("/notes/", response_model=NoteOut)
 async def post_note(note: NoteIn):
     try:
-        print(note.execute_date)
         note_id = db.add_note(note)
 
-        reply = db.get_note_by_id(note_id)
-        return Note(reply[0], reply[1], reply[2], reply[3].strftime(TIME_FORMAT)).__dict__
+        return get_note(note_id)
 
     except Exception:
         raise HTTPException(status_code=500, detail="There was an error posting the note")
@@ -38,32 +41,38 @@ async def get_notes():
 @app.get("/notes/{note_id}", response_model=NoteOut)
 async def get_note_by_id(note_id: int):
     try:
-        reply = db.get_note_by_id(note_id)
-        return Note(reply[0], reply[1], reply[2], reply[3].strftime(TIME_FORMAT)).__dict__
-
+        reply = get_note(note_id)
     except Exception:
         raise HTTPException(status_code=500, detail="There was an error getting the note")
+
+    if not reply:
+        raise HTTPException(status_code=404, detail="There is no note with such id")
+    return reply
 
 
 @app.put("/notes/{note_id}", response_model=NoteOut)
 async def set_note_done(note_id: int):
-    try:
-        db.check_note(note_id)
-        reply = db.get_note_by_id(note_id)
-        return Note(reply[0], reply[1], reply[2], reply[3].strftime(TIME_FORMAT), reply[4]).__dict__
-
-    except Exception:
-        raise HTTPException(status_code=500, detail="There was an error changing the note")
+    if db.check_note_id(note_id):
+        raise HTTPException(status_code=404, detail="There is no note with such id")
+    else:
+        try:
+            db.check_note(note_id)
+            return get_note(note_id)
+        except Exception:
+            raise HTTPException(status_code=500, detail="There was an error changing the not")
 
 
 @app.delete("/notes/{note_id}")
 async def delete_note(note_id: int):
-    try:
-        db.delete_note(note_id)
-        return {'message': f'Note was successfully deleted'}
+    if db.check_note_id(note_id):
+        raise HTTPException(status_code=404, detail="There is no note with such id")
+    else:
+        try:
+            db.delete_note(note_id)
+            return {'message': f'Note was successfully deleted'}
+        except Exception:
+            raise HTTPException(status_code=500, detail="There was an error deleting the note")
 
-    except Exception:
-        raise HTTPException(status_code=500, detail="There was an error deleting the note")
 
 if __name__ == '__main__':
     uvicorn.run(app, host='localhost', port=8000)
